% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Ellipse Area Distribution}
\usage{
area(data)
}
\arguments{
\item{data}{named list of 3 data frames from `breach` function:
`surveyData`, `breachPositionEnsemble` and `breachPositionBestFit`.}
}
\value{
Named list of objects: `ellipse` (sf object of the area of ellipse),
  `fifthPercentileArea` numeric value in metres of 5% percentile area and
  `spotfire_ellipse` (TIBCO Spotfire compatible geometry of ellipse).
}
\description{
Function which produces the ellipse mixing area distribution
}
\examples{
\dontrun{
probability <- probability_non_linear(demo_data)
breach <- breach(probability)
area <- area(breach)
plot(area[["ellipse"]])
area[["fifthPercentileArea"]]
}
}
