% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-taxa.R
\name{assign_taxa}
\alias{assign_taxa}
\title{Assign Taxa}
\usage{
assign_taxa(path = NULL, tax_level = "family", multithread = TRUE)
}
\arguments{
\item{path}{to folder containing fastq files.}

\item{tax_level}{The taxonomic levels being assigned. Default is  "family",
options are c("kingdom", "phylum", "class", "order", "family", "genus",
"species").}

\item{multithread}{Set to FALSE on windows. Only applies to dada2::filterAndTrim
function `https://github.com/benjjneb/dada2/issues/1100`. The rest of the
DADA2 functions will use multithread even on windows.}
}
\value{
dataframe containing assigned fmaily level taxa
}
\description{
Assign family taxa from fastq files.
}
\examples{
\dontrun{
taxa <- assign_taxa(path = demo_path())
}

}
