% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive-stations.R
\name{consecutive_stations}
\alias{consecutive_stations}
\title{Consecutive Stations}
\usage{
consecutive_stations(data)
}
\arguments{
\item{data}{Data frame with survey data}
}
\value{
Data frame with 5 variables.
}
\description{
Check if two consecutive station are at Good status and minimum number of
stations have been sampled.
}
\examples{
\dontrun{
stations <- consecutive_stations(demo_data)
}
}
