% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demo Survey Data}
\format{
A data frame with 30 rows and 15 variables:
\describe{
\item{Survey_date}{Survey_date}
\item{MCFF}{MCFF}
\item{Transect}{Transect}
\item{Station}{Station}
\item{IQI}{IQI}
\item{Easting}{Easting}
\item{Northing}{Northing}
\item{MCFF_Transect}{MCFF_Transect}
\item{Longitude}{Longitude}
\item{Latitude}{Latitude}
\item{Bearing}{Bearing}
\item{Distance}{Distance}
\item{Number of stations per transect}{Number of stations per transect}
\item{WFD status}{WFD status}
\item{MCFF_Transect_Station}{MCFF_Transect_Station}
}
}
\usage{
demo_data
}
\description{
Demo Survey Data
}
\keyword{datasets}
