% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-path.R
\name{demo_path}
\alias{demo_path}
\title{Demo DNA folder}
\usage{
demo_path()
}
\value{
path vector
}
\description{
Path to folder containing DNA and meta data held within `aquaman` package.
}
\examples{
\dontrun{
path <- demo_path()
}
}
