% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-qiime2.R
\name{import_qiime2}
\alias{import_qiime2}
\title{Import taxon data from Qiime2}
\usage{
import_qiime2(path = NULL)
}
\arguments{
\item{path}{File path to the Qiime2 Viewer exported .csv file}
}
\value{
dataframe 682 columns with sample names as row.names.
}
\description{
Import .csv taxonomy file export from Qiime2 Viewer. The taxonomic data is
re-structured to match the format required for the IQI model.
}
\examples{
data <- import_qiime2(system.file("extdat/raw-taxa-data", "S16_TestMOWI.csv", package = "aquaman"))
}
