% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqi.R
\name{iqi}
\alias{iqi}
\title{Predict IQI score}
\usage{
iqi(predictors)
}
\arguments{
\item{predictors}{Data frame of family-level bacteria taxa.}
}
\value{
dataframe of IQI scores
}
\description{
Based on family-level bacteria taxa predictors calculate the benthic invert
IQI outcome.
}
\examples{
\dontrun{
iqi_scores <- iqi(demo_taxa)
}
}
