% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqi-model.R
\docType{data}
\name{iqi_model}
\alias{iqi_model}
\title{IQI model}
\format{
R random forest model object created from the `RRF` package.
}
\usage{
iqi_model
}
\description{
Uses assigned bacteria taxa as predictors to calculate the invertebrate IQI
metric as outcome.
}
\keyword{datasets}
