% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override.R
\name{override}
\alias{override}
\title{Override Distance and Bearing for Transects}
\usage{
override(
  data,
  overrideTransect1 = NA,
  overrideTransect2 = NA,
  overrideTransect3 = NA,
  overrideTransect4 = NA,
  overrideBearing1 = NA,
  overrideBearing2 = NA,
  overrideBearing3 = NA,
  overrideBearing4 = NA
)
}
\arguments{
\item{data}{named list of 4 data frames created by the
`probability_non_linear` function. These are:  `data` (survey), `geoDf`
(distances to good),`geoDfBestFit` (best fit distance) to good and
`hexdfOut` (hexagon heat map).}

\item{overrideTransect1}{Optional override distance Transect 1}

\item{overrideTransect2}{Optional override distance Transect 2}

\item{overrideTransect3}{Optional override distance Transect 3}

\item{overrideTransect4}{Optional override distance Transect 4}

\item{overrideBearing1}{Optional override bearing Transect 1}

\item{overrideBearing2}{Optional override bearing Transect 2}

\item{overrideBearing3}{Optional override bearing Transect 3}

\item{overrideBearing4}{Optional override bearing Transect 4}
}
\value{
area in meters
}
\description{
Override Distance and Bearing for Transects
}
\examples{
\dontrun{
probability <- probability_non_linear(demo_iqi)
probability <- override(probability)
}
}
