% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-non-linear.R
\name{probability_non_linear}
\alias{probability_non_linear}
\title{Probability Non-linear}
\usage{
probability_non_linear(data)
}
\arguments{
\item{data}{Data frame with survey data}
}
\value{
list containing four named data frames: data, geoDf, geoDfBestFit and
  hexdfOut.
}
\description{
Probability Non-linear
}
\examples{
\dontrun{
probability <- probability_non_linear(demo_iqi)
}
}
