% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiime2-taxa-assign.R
\name{qiime2_taxa_assign}
\alias{qiime2_taxa_assign}
\title{Qiime2 taxa assign}
\usage{
qiime2_taxa_assign(path = NULL)
}
\arguments{
\item{path}{Path to directory containing .fasta files}
}
\value{
dataframe
}
\description{
Only on Linux, Mac or WSL
}
\examples{
\dontrun{
path <- "my_data_folder/"
qiime2_taxa_assign(path)
}
}
