% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Assess survey data}
\usage{
assess(data)
}
\arguments{
\item{data}{Data frame with survey data}
}
\value{
area in meters
}
\description{
Assess if IQI survey data breaches mixing zone.
}
\examples{
\dontrun{
area <- assess(demo_iqi)
}
}
