% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-taxa.R
\name{assign_taxa}
\alias{assign_taxa}
\title{Assign Taxa}
\usage{
assign_taxa(path = NULL, tax_level = "family", multithread = FALSE)
}
\arguments{
\item{path}{to folder containing folder(s) of fastq files. Assuming
DNA foldernames contain `001` and filenames have format
`SAMPLENAME_XXX.fastq` - as per normal Illumina naming.}

\item{tax_level}{The taxonomic levels being assigned. Default is  "family",
options are c("kingdom", "phylum", "class", "order", "family", "genus",
"species").}

\item{multithread}{Set to FALSE on windows. Only applies to
dada2::filterAndTrim
function `https://github.com/benjjneb/dada2/issues/1100`. The rest of the
DADA2 functions will use multithread even on windows.}
}
\value{
dataframe containing assigned fmaily level taxa
}
\description{
WORK IN PROGRESS - DO NOT USE IN PRODUCTION - Assign family taxa from .fastq
files. Will prompt you to download taxonomic reference file on first use.
}
\examples{
\dontrun{
taxa <- assign_taxa(path = demo_path())
}

}
