% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiime2_taxa_assign.R
\name{qiime2_taxa_assign}
\alias{qiime2_taxa_assign}
\title{Qiime2 taxa assign}
\usage{
qiime2_taxa_assign(path = NULL)
}
\arguments{
\item{path}{Path to directory containing .fasta files}
}
\description{
Only on Linux, Mac or WSL
}
\examples{
path = "OneDrive - Scottish Environment Protection Agency/Reports/Fish Farm/operating_iqi_surveys/ScreeningTool_SequenceData"
qiime2_taxa_assign
}
